reorganized server endpoints for consistency
