+++ backend/new API.md
@@ -1,0 +1,21 @@
+# /Auth
+## /Auth/Login
+## /Auth/Logout
+## /Auth/Google
+## /Auth/Google/Redirect
+
+# /Profile
+## /Profile/
+
+# /Pay
+## /Pay/Charge
+## /Pay/Cancel
+
+# /User
+## /User/Info/:ID
+## /User/Info/Subscribe/:ID
+## /User/Info/Unsubscribe/:ID
+
+# /Items
+
+# /Outfits
